pvr/epg: refactor EPG updates, resulting in much shorter locking durations. fixes non-responsive UI during EPG updates

+++ xbmc/pvr/PVRGUIInfo.cpp
@@ -102,7 +102,7 @@ bool CPVRGUIInfo::AddonInfoToggle(void)
 
   if (CTimeUtils::GetTimeMS() - m_iAddonInfoToggleStart > INFO_TOGGLE_TIME)
   {
-    if (++m_iAddonInfoToggleCurrent > m_iActiveClients - 1)
+    if (((int) ++m_iAddonInfoToggleCurrent) > m_iActiveClients - 1)
       m_iAddonInfoToggleCurrent = 0;
 
     return true;
